Split views.py in several files;delete models.py,views.py
