+++ lowfat/views/__init__.py
@@ -1,0 +1,4 @@
+from .blog import *
+from .claimant import *
+from .expense import *
+from .fund import *
